fix: menu-cardコンポーネントの修正
