+++ Heritage-View/src/app/components/menu-card/menu-card.component.html
@@ -1,0 +1,7 @@
+
+<mat-card class="card">
+    <div class="image-trim">
+        <img mat-card-image src={{card.src}} alt="world image" class="card-img">
+    </div>
+    <mat-card-title>{{card.title}}</mat-card-title>
+</mat-card>
