+++ Heritage-View/src/app/components/menu-card/menu-card.component.scss
@@ -17,25 +17,3 @@
         object-fit: cover;
     }
 }
-
-.wrap {
-    display: flex;
-    flex-wrap: wrap;
-}
-
-.item {
-    padding: 10px;
-    width: calc((100% - 30px * 2) / 3);
-}
-
-@media not all and (min-width: 960px) {
-    .item {
-        width: calc((100% - 30px * 2) / 2)
-    }
-}
-
-@media not all and (min-width: 600px) {
-    .item {
-        width: 100%
-    }
-}
