--- Heritage-View/src/app/components/menu-cards/menu-cards.component.html
@@ -1,12 +1,0 @@
-<div class="wrap">
-    <div class="item" *ngFor="let card of cards">
-        <button mat-button>
-            <mat-card class="card">
-                <div class="image-trim">
-                    <img mat-card-image src={{card.src}} alt="world image" class="card-img">
-                </div>
-                <mat-card-title>{{card.title}}</mat-card-title>
-            </mat-card>
-        </button>
-    </div>
-</div>
