+++ Heritage-View/src/app/pages/top/top.component.html
@@ -1,5 +1,5 @@
-<div class="content">
-    <div class="top-container">
+<div class="container">
+    <div class="content">
             <app-difficulty-set class="margin-left"></app-difficulty-set>
 
             <app-primary-button [text]="explanationButtonText" class="explanation-button"></app-primary-button>
@@ -7,6 +7,12 @@
     
     <p class="margin-left explanatory">選択された地域の世界遺産が計５問出題されます</p>
 
-    <app-menu-cards></app-menu-cards>
+    <div class="menu-card-content">
+        <div class="wrap" *ngFor="let card of cards">
+            <button mat-button>
+                <app-menu-card [card]=card></app-menu-card>
+            </button>
+        </div>
+    </div>
 </div>
